Fixes for new firmware

+++ silicon_on_demand/run_job.py
@@ -8,15 +8,15 @@ import tempfile
 import optparse
 import signal
 
-SB_01 = { "uart": "/dev/ttyUSB1",
-          "baud": "115200",
-          "chain_uboot": "/home/cgray/uboot-tftp.mpg.gz",
+SB_01 = { "uart": "/dev/ttyUSB4",
+          "baud": "1000000",
+#          "chain_uboot": "/home/cgray/uboot-tftp.mpg.gz",
           "serverip": "172.17.1.1",
           "boardip": "172.17.1.2",
           "iface": "enp1s0f1"}
 
 SB_02 = { "uart": "/dev/ttyUSB0",
-          "baud": "115200",
+          "baud": "1000000",
           "serverip": "172.16.1.1",
           "boardip": "172.16.1.2",
           "iface": "enp1s0f0"}
@@ -96,7 +96,7 @@ funos_crashing:
   print "FunOS crashed, but didn't see platform_halt"
 
 halt_timeout:
-  print "Timeout waiting for FunOS to platform_halt"
+  print "Timeout waiting for FunOS to platform halt"
   goto out
 
 no_unzip:
@@ -110,13 +110,15 @@ no_uboot:
 
 out:
   sleep 1
+  print "boot.script: killing minicom"
   ! cat %s | xargs kill -HUP 
 """
 
 SCRIPT_CHAIN = """
+  send ""
   expect {{
   	 "f1 #" break
-  	 timeout 20  goto no_uboot
+  	 timeout 40  goto no_uboot
   }}
 
   send "loadx"
@@ -139,17 +141,22 @@ SCRIPT_TFTP = """
   send ""
 
 # now wait for new u-boot
+#  expect {{
+#  	 "Type 'noautoboot' to disable autoboot"
+#  	 timeout 20 goto no_uboot
+#  }}
+
   expect {{
-  	 "Type 'noautoboot' to disable autoboot"
-  	 timeout 20 goto no_uboot
+  	 "f1 #"
+  	 timeout 20 exit
   }}
 
 chain_bypass:
 
   timeout {global_timeout}
-  send "noautoboot"
+#  send "noautoboot"
 
-  print "\\nboot.script: autoboot cancelled"
+#  print "\\nboot.script: autoboot cancelled"
 
   print "\\nboot.script: bringing up the network"
   send "lfw"
@@ -216,13 +223,13 @@ chain_bypass:
   send "tftpboot 0xffffffff91000000 {serverip}:{funos} ; unzip 0xFFFFFFFF91000000 0xa800000020000000 ; bootelf -p 0xa800000020000000"
   print "\\n"
 
-  print "boot.script: waiting for FunOS to platform_halt now"
+  print "boot.script: waiting for FunOS to platform halt now"
 
   # if we see FunOS, expect bootstrap
-  expect {{
-      "Welcome to FunOS" break
-      timeout 10 goto halt_wait
-  }}
+  #expect {{
+  #    "Welcome to FunOS" break
+  #    timeout 10 goto halt_wait
+  #}}
 
   expect {{
       "sending bootstrap WU" break
@@ -246,7 +253,7 @@ halt_wait:
   goto out
 
 halt_timeout:
-  print "Timeout waiting for FunOS to platform_halt"
+  print "boot.script: Timeout waiting for FunOS to platform halt"
   goto out
 no_uboot:
   print "boot.script: no uboot"
@@ -262,6 +269,7 @@ eep2:
   goto out
 eep3:
   print "boot.script: eep3"
+  ! touch restart.fail
   print "\\n"
   goto out
 eep4:
@@ -376,6 +384,7 @@ else:
 
     script_name = "%s/boot.script" % path
     log_name = "%s/minicom-log" % path
+    exit_name = "%s/exitlog" % path
     pid_name = "%s/minicom.pid" % path
     dun_name = "%s/job.done" % path
 
@@ -412,8 +421,14 @@ else:
         d['serverip'] = board['serverip']
         d['boardip'] = board['boardip']
         d['iface'] = board['iface']
+        
         if (board.get("chain_uboot") is not None):
             d['chain_uboot'] = board["chain_uboot"]
+            
+        if (options.uboot is not None):
+            d['chain_uboot'] = options.uboot
+            
+        if (d.get("chain_uboot") is not None):
             script = SCRIPT_CHAIN.format(**d)
             script += SCRIPT_TFTP.format(**d)
         else:
@@ -433,11 +448,12 @@ else:
 
     device = boards[options.board]["uart"]
     baud = boards[options.board]["baud"]
-    cmd = "minicom -D %s -b %s -S %s -w -C %s" % (device, baud, script_name, log_name)
+    os.system("echo '\nStarting job on board %s' >> %s" % (options.board, log_name))
+    cmd = "/usr/bin/minicom -D %s -b %s -S %s -w -C %s" % (device, baud, script_name, log_name)
 
     # fork minicom to get its pid for later
     print "Forking minicom..."
-    p = subprocess.Popen(cmd, shell=True)
+    p = subprocess.Popen(cmd.split())
 
     fl = open(pid_name, "w")
     fl.write("%s" % p.pid)
@@ -453,12 +469,24 @@ else:
         if (global_timeout is not None):
             if ((time.time() - t0) > global_timeout):
                 print "global timeout reached"
+                os.system("echo 'run_job: global timeout reached' >> %s" % (exit_name))
                 do_kill = True
 
         if (os.path.exists(dun_name)):
             print "someone requested a kill, killing minicom"
+            os.system("echo 'run_job: kill requested' >> %s" % (exit_name))
             do_kill = True
 
+        # double check for platform_halt because runscript is terrible
+        out = subprocess.Popen(['tail', '-3', log_name], 
+                               stdout=subprocess.PIPE, 
+                               stderr=subprocess.STDOUT)
+        stdout,stderr = out.communicate()
+        if "platform_halt" in stdout:
+            print "platform halt detected, killing minicom"
+            os.system("echo 'run_job: platform halt detected' >> %s" % (exit_name))
+            do_kill = True
+            
         if (do_kill):
             killcmd = "kill -HUP %s" % p.pid
             print "killing with '%s'" % killcmd
@@ -467,13 +495,29 @@ else:
             p.wait()
             break
             
+    if (p.returncode is not None):
+        print "run_job: minicom exited"
+        cmd = "echo 'run_job: minicom died of natural causes (%s)' >> %s" % (p.returncode, exit_name)
+        print cmd
+        os.system(cmd)
+    else:
+        print "run_job: we killed minicom"
+        cmd = "echo 'run_job: minicom died of unnatural causes (%s)' >> %s" % (p.returncode, exit_name)
+        print cmd
+        os.system(cmd)
 
+    #for i in range(30):
+    #    print "run_job: pause"
+    #    do_sleep(1)
+        
     print "Minicom is dead. long live minicom, but fixing its terminal settings..."
     do_sleep(2)
     if (not options.no_terminal_reset):
         os.system("reset")
     print "log file at %s.txt" % log_name
     os.system("cat %s | grep -v 'Xmodem sectors' > %s.txt" % (log_name, log_name))
+    os.system("cat %s >> %s.txt" % (exit_name, log_name))
+    os.system("echo 'Job completed on board %s' >> %s.txt" % (options.board, log_name))
     os.system("cat %s.txt | tail -20" % log_name)
 
     print "Cleaning my own process group"
